Started on reading new format, displaying something...
Converted biomes to use int instead of bytes

+++ src/org/jmc/BlockInfo.java
@@ -88,7 +88,7 @@ public class BlockInfo
 	 * @param data Block data
 	 * @return Block color
 	 */
-	public Color getPreviewColor(byte data, byte biome)
+	public Color getPreviewColor(byte data, int biome)
 	{
 		String[] mtlNames = getMaterials().get(data,biome);
 		if (mtlNames == null || mtlNames.length == 0)
