+++ src/org/jmc/models/Banner.java
@@ -74,7 +74,7 @@ public class Banner extends BlockModel {
     }
 
     @Override
-    public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, byte biome) {
+    public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, int biome) {
 
         //Log.info("Banner ***************************");
