+++ src/org/jmc/models/CocoaPlant.java
@@ -14,7 +14,7 @@ public class CocoaPlant extends BlockModel
 {
 
 	@Override
-	protected String[] getMtlSides(byte data, byte biome)
+	protected String[] getMtlSides(byte data, int biome)
 	{
 		int growth = (data >> 2) & 3;
 		
@@ -32,7 +32,7 @@ public class CocoaPlant extends BlockModel
 
 	
 	@Override
-	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, byte biome)
+	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, int biome)
 	{
 		int dir = data & 3;
 		int growth = (data >> 2) & 3;
