+++ src/org/jmc/models/Cross.java
@@ -13,7 +13,7 @@ public class Cross extends BlockModel
 {
 
 	@Override
-	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, byte biome)
+	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, int biome)
 	{
 		Transform move = new Transform();
 		move.translate(x, y, z);
