+++ src/org/jmc/models/DirtGrass.java
@@ -10,7 +10,7 @@ import org.jmc.threading.ThreadChunkDeligate;
 public class DirtGrass extends BlockModel
 {
 	
-	protected String[] getMtlSides(byte data, byte biome, boolean snow)
+	protected String[] getMtlSides(byte data, int biome, boolean snow)
 	{
 		String[] abbrMtls = materials.get(data,biome);
 		
@@ -27,7 +27,7 @@ public class DirtGrass extends BlockModel
 	
 
 	@Override
-	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, byte biome)
+	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, int biome)
 	{
 		boolean snow = chunks.getBlockID(x, y+1, z) == 78;
