+++ src/org/jmc/models/Farmland.java
@@ -12,7 +12,7 @@ public class Farmland extends BlockModel
 {
 	
 	@Override
-	protected String[] getMtlSides(byte data, byte biome)
+	protected String[] getMtlSides(byte data, int biome)
 	{
 		String[] abbrMtls = materials.get(data,biome);
 
@@ -28,7 +28,7 @@ public class Farmland extends BlockModel
 	
 
 	@Override
-	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, byte biome)
+	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, int biome)
 	{
 		boolean[] drawSides = drawSides(chunks, x, y, z);
