+++ src/org/jmc/models/Quartz.java
@@ -12,7 +12,7 @@ public class Quartz extends BlockModel
 {
 
 	@Override
-	protected String[] getMtlSides(byte data, byte biome)
+	protected String[] getMtlSides(byte data, int biome)
 	{
 		String[] mat = materials.get(data, biome);
 		
@@ -43,7 +43,7 @@ public class Quartz extends BlockModel
 
 
 	@Override
-	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, byte biome)
+	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, int biome)
 	{
 		addBox(obj,
 				x - 0.5f, y - 0.5f, z - 0.5f,
