+++ src/org/jmc/models/RedstoneWire.java
@@ -22,7 +22,7 @@ public class RedstoneWire extends BlockModel
 	}
 	
 	@Override
-	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, byte biome)
+	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, int biome)
 	{
 		boolean on = (data & 15) > 0;
 		String mtlCross = on ? materials.get(data,biome)[0] : materials.get(data,biome)[2];
