+++ src/org/jmc/models/Torch.java
@@ -14,7 +14,7 @@ public class Torch extends BlockModel
 {
 
 	@Override
-	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, byte biome)
+	public void addModel(ChunkProcessor obj, ThreadChunkDeligate chunks, int x, int y, int z, byte data, int biome)
 	{
 		String[] mtls = materials.get(data,biome);
